refactor: Update main.py to ProductResource.py
Update main.py to ProductResource.py and refactor code structure
